Converted all relationships to one2many type with backrefs
Aside from special cases there's no need to explicitly create relationship
attributes on both sides. And for one2many is much more convenient if the
relationship is created on the one side, because on average that means fewer
relationship objects per model. They are also easier to track that way.

+ fixed some ordering that was broken due to incorrect usage of backrefs.