Rename .follow-button class to workaround over-eager adblockers
Renames the .follow-button class to .follow-mod-button (and .unfollow-button to .unfollow-mod-button to stay consistent).
Several times users reported missing download or follow buttons.
While I can't confirm it for the download button, I can reproduce it for the follow button thanks to a user reporting it together with the ad-blocker they are using.

It seems to be the "Block Tracking through Social-Media-Symbols" option in Adblock Plus.
It is based on the "Fanboy's Social Blocking List", which includes the filter ##.follow-button .

Renaming the class to .follow-mod-button resolves the problem.