Move piwik loading stuff into own static JS file
This reduces the size of all HTML pages by around 33KB.
Which is aound 1/3 to 2/3 of their size.

Being in its own static file, this will be offloaded by the webserver,
and the reverse proxy and clients will be able to cache it.

This should reduce load on Flask/gunicorn by quite a lot.